Revert @kyleshevlin / #8,#9.
I had a change of heart and decided to roll back this functionality, as I couldn't reconcile bundle size / performance with theusefulness of the feature. For example, we had to add code to check whether a prefix was being used or not, because that case represents the exception to the rule.

Not on #9, but I also want to add I don't want to convert the number to a string, since the user, most likely using a VDOM is going to turn numbers to strings anyway.

    cw(1) // => 1

I also don't think classwrap should throw  and would prefer to pass on the responsibility to the user / VDOM